fix warnings

+++ Eclipse/src/Editor/Windows/AssetBrowser/AssetBrowser.cpp
@@ -355,7 +355,7 @@ namespace Eclipse
 
 			TextureComponent icon = dirEntry.is_directory() ? FolderIcon : sprite;
 
-			ImGui::ImageButton((void*)Graphics::FindTextures(icon.textureRef).GetHandle(),
+			ImGui::ImageButton((void*)(intptr_t)Graphics::FindTextures(icon.textureRef).GetHandle(),
 				buttonSize,
 				{ 1,0 },
 				{ 2,1 });
@@ -482,7 +482,7 @@ namespace Eclipse
 		// If element was found
 		if (it != container.end())
 		{
-			pos = it - container.begin();
+			pos = static_cast<int>(it - container.begin());
 		}
 	}
 
@@ -667,7 +667,7 @@ namespace Eclipse
 				{
 					TextureComponent icon = std::filesystem::is_directory(pair2) ? FolderIcon : sprite;
 
-					ImGui::ImageButton((void*)Graphics::FindTextures(icon.textureRef).GetHandle(),
+					ImGui::ImageButton((void*)(intptr_t)Graphics::FindTextures(icon.textureRef).GetHandle(),
 						buttonSize,
 						{ 1,0 },
 						{ 2,1 });
@@ -711,7 +711,7 @@ namespace Eclipse
 				{
 					TextureComponent icon = std::filesystem::is_directory(tempPath) ? FolderIcon : sprite;
 
-					ImGui::ImageButton((void*)Graphics::FindTextures(icon.textureRef).GetHandle(),
+					ImGui::ImageButton((void*)(intptr_t)Graphics::FindTextures(icon.textureRef).GetHandle(),
 						buttonSize,
 						{ 1,0 },
 						{ 2,1 });
